backend/requirements.txt
Django==5.2
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.0.1
pandas==2.1.1
numpy==1.24.3
scipy==1.11.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
reportlab==4.0.4
openpyxl==3.1.2
